+++ Assets/Plugins/GameService/Utils/RealTimeUtil/Classes/Observers/RealtimeRigidbody2DObserver.cs
@@ -65,7 +65,7 @@ namespace Plugins.GameService.Utils.RealTimeUtil.Classes.Observers
         
         public void FixedUpdate()
         {
-            _rBody.position = Vector3.MoveTowards(_rBody.position, _mNetworkPosition, _mDistance * (1.0f / GsLiveRealtime.SerializationRate));
+            _rBody.position = Vector2.MoveTowards(_rBody.position, _mNetworkPosition, _mDistance * (1.0f / GsLiveRealtime.SerializationRate));
             _rBody.rotation = Mathf.MoveTowards(_rBody.rotation, _mNetworkRotation, _mAngle * (1.0f / GsLiveRealtime.SerializationRate));
         }