+++ Assets/Plugins/GameService/Utils/RealTimeUtil/Models/UnitySerializerModels/Primitives/ٰVector3Serializer.cs
@@ -30,18 +30,13 @@ namespace Plugins.GameService.Utils.RealTimeUtil.Models.UnitySerializerModels.Pr
     {
         protected override void WriteObject(Vector3 obj, GsWriteStream writeStream)
         {
-            writeStream.WriteNext(obj.x);
-            writeStream.WriteNext(obj.y);
-            writeStream.WriteNext(obj.z);
+            writeStream.WriteNext(new[]{obj.x,obj.y,obj.z});
         }
 
         protected override Vector3 ReadObject(GsReadStream readStream)
         {
-            var x = (float) readStream.ReadNext();
-            var y = (float) readStream.ReadNext();
-            var z = (float) readStream.ReadNext();
-            
-            return new Vector3(x,y,z);
+            var data = (float[]) readStream.ReadNext();
+            return new Vector3(data[0],data[1],data[2]);
         }
     }
 }